update install script to have options; add confirm shell script

+++ install.sh
@@ -1,22 +1,35 @@
 #!/bin/bash
-echo Home directory: $HOME
-if [ ! -d ~/.git ]; then
-  echo Cloning...
+source confirm.sh
+if [ ! -d ~/.git ] && confirm "Install dotfiles and configs"; then
+  exit
   git clone https://github.com/joshuajwilborn/vim.git ~/vim
   echo Moving files...
   rsync -a ~/vim/ ~/
   echo Cleaning up...
   rm -rf ~/vim
-	git remote remove origin
-	git remote add origin git@github.com:JoshuaJWilborn/vim.git
+  git remote remove origin
+  git remote add origin git@github.com:JoshuaJWilborn/vim.git
 else
   echo Skipping git clone...
 fi 
 
-sudo pacman --needed --noconfirm -S $(cat ~/.packages)
-sudo pip install powerline-status undervolt
+if [ $(command -v pacman) ] && confirm "Install pacman and pip and pip packages:"; then
+  echo "Installing pacman packages..."
+  sudo pacman --needed --noconfirm -S $(cat ~/.packages)
 
-if [ ! $(command -v fzf) ]; then
+  echo "Installing pip packages"
+  sudo pip install powerline-status undervolt
+
+  echo "Installing TLP services"
+  sudo echo TLP_ENABLE=1 | sudo tee /usr/share/defaults/tlp
+  sudo systemctl enable tlp.service
+  sudo systemctl enable tlp-sleep.service
+  sudo systemctl enable NetworkManager-dispatcher.service
+  sudo systemctl mask systemd-rfkill.service
+  sudo systemctl mask systemd-rfkill.socket
+fi
+
+if [ ! $(command -v fzf) ] && confirm "Install FZF (fuzzy file finder):"; then
   git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
   ~/.fzf/install
   rm -rf ~/.fzf
@@ -24,31 +37,22 @@ fi
 
 vim -E -s -u "~/.vimrc" +PlugInstall +CoCInstall +qall
 
-sudo echo TLP_ENABLE=1 | sudo tee /usr/share/defaults/tlp
-sudo systemctl enable tlp.service
-sudo systemctl enable tlp-sleep.service 
-sudo systemctl enable NetworkManager-dispatcher.service
-sudo systemctl mask systemd-rfkill.service
-sudo systemctl mask systemd-rfkill.socket
-
-# install linux brew
-#if [ ! $(command -v brew) ]; then
-#  sh -c "$(curl -fsSL https://raw.githubusercontent.com/Linuxbrew/install/master/install.sh)"
-#fi
-
-
-sudo tee -a /etc/systemd/system/undervolt.service <<- END
+if [ $(command -v undervolt) ] && confirm "Install undervolt service(-135mv): "; then
+  sudo tee -a /etc/systemd/system/undervolt.service <<- END
   [Unit]
   Description=undervolt
   [Service]
   Type=oneshot
   ExecStart=`which undervolt` -v --core -135 --cache -135 --gpu -135 --analogio -135 --uncore -135
 END
-sudo tee -a /etc/systemd/system/undervolt-reset.service <<- END
+  sudo tee -a /etc/systemd/system/undervolt-reset.service <<- END
   [Unit]
   Description=undervolt
   [Service]
   Type=oneshot
   ExecStart=`which undervolt` -v --core -0 --cache -0 --gpu -0 --analogio -0 --uncore -0
 END
-source ~/.bashrc
+  echo "Undervolt by typing 'undervolt' and reset by calling 'undervolt-reset'"
+fi
+
+
